:recycle: refactor: extract MapContextMenu from MapView
Resolves #789 (Phase 1)

Extracted complex context menu logic from MapView into a separate MapContextMenu component to reduce file size and simplify rendering logic. Added MAP_VIEW_REFACTOR.md to track overall progress.
